+++ eslint-plugin/tests/lib/rules/avoid-css-animations.js
@@ -1,0 +1,80 @@
+/**
+ * Copyright (C) 2023 Green Code Initiative
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+"use strict";
+
+//------------------------------------------------------------------------------
+// Requirements
+//------------------------------------------------------------------------------
+
+const rule = require("../../../lib/rules/avoid-css-animations");
+const RuleTester = require("eslint").RuleTester;
+
+//------------------------------------------------------------------------------
+// Tests
+//------------------------------------------------------------------------------
+
+const ruleTester = new RuleTester({
+  parserOptions: {
+    ecmaVersion: 2021,
+    sourceType: "module",
+    ecmaFeatures: {
+      jsx: true,
+    },
+  },
+});
+
+ruleTester.run("avoid-css-animations", rule, {
+  valid: [
+    `
+    import React from 'react';
+    import './styles.css'; // External CSS file
+
+    const MyComponent = () => {
+      return <div className="my-class">This content is styled using an external CSS file.</div>;
+    };
+
+    export default MyComponent;
+    `,
+  ],
+
+  invalid: [
+    {
+      code: "<div style={{ transition: 'width 2s' }} />",
+      errors: [
+        {
+          messageId: "AvoidCSSAnimations",
+          data: {
+            attribute: "transition",
+          },
+          type: "Property",
+        },
+      ],
+    },
+    {
+      code: "<div style={{ animationName: 'example', animationDuration: '4s' }} />",
+      errors: [
+        {
+          messageId: "AvoidCSSAnimations",
+          data: {
+            attribute: "animationName",
+          },
+          type: "Property",
+        },
+      ],
+    },
+  ],
+});